backend/requirements.txt
# Deep Learning
torch==2.6.0
torchvision==0.21.0

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
albumentations>=1.3.0

# Metrics and Visualization
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/Chatbot
openai>=1.3.0
transformers>=4.35.0

# Testing
pytest>=7.4.0
httpx>=0.25.0